requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.0
Pillow==10.2.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
qrcode==7.4.2
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0